removing the file describing my textmate project as is different on different computers

+++ vtkCenterOfMassFilter.cxx
@@ -40,7 +40,7 @@ int vtkCenterOfMassFilter::FillInputPortInformation(int, vtkInformation* info)
   return 1;
 }
 
-double* vtkCenterOfMassFilter::CalculateWeightedMass(double& mass,\
+float* vtkCenterOfMassFilter::CalculateWeightedMass(double& mass,\
 																double* point)
 {
 	double* weightedMass = new double[3];
